+++ app/src/actions/types.js
@@ -1,2 +1,14 @@
-export const SHOW_NOTIFICATION = 'SHOW_NOTIFICATION'
-export const APP_CHANGE_ONLINE_STATUS = 'APP_CHANGE_ONLINE_STATUS'
+// app
+export const SHOW_NOTIFICATION = '[app] SHOW_NOTIFICATION'
+export const APP_CHANGE_ONLINE_STATUS = '[app] CHANGE_ONLINE_STATUS'
+export const APP_LAUNCHED = '[app] LAUNCHED'
+export const APP_SHOW_ONBOARDING = '[app] SHOW_ONBOARDING'
+export const APP_USER_HAS_INACTIVE_TOKEN = '[app] USER_HAS_INACTIVE_TOKEN'
+// user
+export const CHANGE_USER_DATA = '[user] CHANGE_USER_DATA'
+export const USER_LOGIN_FAILED = '[user] LOGIN_FAILED'
+export const USER_LOGIN_SUCCESS = '[user] LOGIN_SUCCESS'
+export const USER_LOGIN_PENDING = '[user] LOGIN_PENDING'
+export const USER_EMAIL_LOGIN = '[user] EMAIL_LOGIN'
+export const USER_ETH_LOGIN = '[user] ETH_LOGIN'
+export const USER_LOGOUT_SUCCESS = '[user] LOGOUT_SUCCESS'