+++ app/src/utils/utils.js
@@ -15,11 +15,10 @@ const calculateBTCMarketcap = ({marketcap, priceUsd, priceBtc}) => {
 }
 
 const getOrigin = () => {
-  const origin = (window.env || {}).WEBSITE_URL || ''
-  if (process.env.production) {
-    return window.location.origin
+  if (process.env.NODE_ENV === 'development') {
+    return process.env.REACT_APP_WEBSITE_URL || window.location.origin
   }
-  return origin
+  return window.location.origin
 }
 
 const sanitizeMediumDraftHtml = (html) => sanitizeHtml(html,